renamed SNAPExportTools to blueUtils and updated all references. This is a prelude to expanding its scope
